rename module to `github_indices` and switch to using llama index
